requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
python-dateutil==2.5.2
pytz==2016.3
six==1.10.0
